requirements.txt
streamlit
google-genai
num2words
pandas
fpdf2
python-dotenv
num2words